commit 11/07/2024

+++ asset-service/src/main/java/com/duc/asset_service/controller/AssetController.java
@@ -6,6 +6,7 @@ import com.duc.asset_service.model.Asset;
 import com.duc.asset_service.service.AssetService;
 import com.duc.asset_service.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -18,24 +19,32 @@ import java.util.List;
 public class AssetController {
     private final AssetService assetService;
     private final UserService userService;
+    @Value("${internal.service.token}")
+    private String internalServiceToken;
 
     @PostMapping
-    public ResponseEntity<Asset> createAsset(@RequestHeader("Authorization") String jwt, @RequestBody CreateAssetRequest request) throws Exception {
-        UserDTO user = userService.getUserProfile(jwt);
-        Asset asset = assetService.createAsset(user.getId(), request.getCoinId(), request.getQuantity());
+    public ResponseEntity<Asset> createAsset(@RequestHeader("Internal-Service-Token") String jwt, @RequestBody CreateAssetRequest request) throws Exception {
+        if (!internalServiceToken.equals(jwt)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        Asset asset = assetService.createAsset(request.getUserId(), request.getCoinId(), request.getQuantity());
         return new ResponseEntity<>(asset, HttpStatus.OK);
     }
 
     @PutMapping("/{assetId}")
-    public ResponseEntity<Asset> updateAsset(@RequestHeader("Authorization") String jwt, @PathVariable Long assetId, @RequestParam("quantity") double quantity) throws Exception {
-        UserDTO user = userService.getUserProfile(jwt);
+    public ResponseEntity<Asset> updateAsset(@RequestHeader("Internal-Service-Token") String jwt, @PathVariable Long assetId, @RequestParam("quantity") double quantity) throws Exception {
+        if (!internalServiceToken.equals(jwt)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         Asset asset = assetService.updateAsset(assetId, quantity);
         return new ResponseEntity<>(asset, HttpStatus.OK);
     }
 
     @DeleteMapping("/{assetId}")
-    public ResponseEntity<String> deleteAsset(@RequestHeader("Authorization") String jwt, @PathVariable Long assetId) throws Exception {
-        UserDTO user = userService.getUserProfile(jwt);
+    public ResponseEntity<String> deleteAsset(@RequestHeader("Internal-Service-Token") String jwt, @PathVariable Long assetId) throws Exception {
+        if (!internalServiceToken.equals(jwt)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         assetService.deleteAsset(assetId);
         return new ResponseEntity<>("Delete asset successfully.", HttpStatus.OK);
     }
